+++ src/SMAPI.Installer/Program.cs
@@ -1,7 +1,10 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using System.IO.Compression;
 using System.Reflection;
+using StardewModdingAPI.Internal;
+using StardewModdingAPI.Toolkit.Utilities;
 
 namespace StardewModdingApi.Installer
 {
@@ -11,9 +14,15 @@ namespace StardewModdingApi.Installer
         /*********
         ** Properties
         *********/
-        /// <summary>The absolute path to search for referenced assemblies.</summary>
+        /// <summary>The absolute path of the installer folder.</summary>
         [SuppressMessage("ReSharper", "AssignNullToNotNullAttribute", Justification = "The assembly location is never null in this context.")]
-        private static readonly string DllSearchPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "smapi-internal");
+        private static readonly string InstallerPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+
+        /// <summary>The absolute path of the folder containing the unzipped installer files.</summary>
+        private static readonly string ExtractedBundlePath = Path.Combine(Path.GetTempPath(), $"SMAPI-installer-{Guid.NewGuid():N}");
+
+        /// <summary>The absolute path for referenced assemblies.</summary>
+        private static readonly string InternalFilesPath = Path.Combine(Program.ExtractedBundlePath, "smapi-internal");
 
         /*********
         ** Public methods
@@ -22,11 +31,26 @@ namespace StardewModdingApi.Installer
         /// <param name="args">The command line arguments.</param>
         public static void Main(string[] args)
         {
+            // find install bundle
+            PlatformID platform = Environment.OSVersion.Platform;
+            FileInfo zipFile = new FileInfo(Path.Combine(Program.InstallerPath, $"{(platform == PlatformID.Win32NT ? "windows" : "mono")}.dat"));
+            if (!zipFile.Exists)
+            {
+                Console.WriteLine($"Oops! Some of the installer files are missing; try redownloading the installer. (Missing file: {zipFile.FullName})");
+                Console.ReadLine();
+                return;
+            }
+
+            // unzip bundle into temp folder
+            DirectoryInfo bundleDir = new DirectoryInfo(Program.ExtractedBundlePath);
+            Console.WriteLine("Extracting install files...");
+            ZipFile.ExtractToDirectory(zipFile.FullName, bundleDir.FullName);
+
             // set up assembly resolution
             AppDomain.CurrentDomain.AssemblyResolve += Program.CurrentDomain_AssemblyResolve;
 
             // launch installer
-            var installer = new InteractiveInstaller();
+            var installer = new InteractiveInstaller(bundleDir.FullName);
             installer.Run(args);
         }
 
@@ -41,7 +65,7 @@ namespace StardewModdingApi.Installer
             try
             {
                 AssemblyName name = new AssemblyName(e.Name);
-                foreach (FileInfo dll in new DirectoryInfo(Program.DllSearchPath).EnumerateFiles("*.dll"))
+                foreach (FileInfo dll in new DirectoryInfo(Program.InternalFilesPath).EnumerateFiles("*.dll"))
                 {
                     if (name.Name.Equals(AssemblyName.GetAssemblyName(dll.FullName).Name, StringComparison.InvariantCultureIgnoreCase))
                         return Assembly.LoadFrom(dll.FullName);